Explain RowSetFinishing
