+++ src/coord/coord.rs
@@ -36,11 +36,13 @@ use dataflow::logging::materialized::MaterializedEvent;
 use dataflow::{SequencedCommand, WorkerFeedback, WorkerFeedbackWithMeta};
 use dataflow_types::logging::LoggingConfig;
 use dataflow_types::{
-    DataflowDesc, IndexDesc, KafkaSinkConnector, PeekResponse, PeekWhen, RowSetFinishing,
-    SinkConnector, TailSinkConnector, Timestamp, Update,
+    DataflowDesc, IndexDesc, KafkaSinkConnector, PeekResponse, PeekWhen, SinkConnector,
+    TailSinkConnector, Timestamp, Update,
 };
 use expr::transform::Optimizer;
-use expr::{EvalEnv, GlobalId, Id, IdHumanizer, RelationExpr, ScalarExpr, SourceInstanceId};
+use expr::{
+    EvalEnv, GlobalId, Id, IdHumanizer, RelationExpr, RowSetFinishing, ScalarExpr, SourceInstanceId,
+};
 use ore::collections::CollectionExt;
 use ore::thread::JoinHandleExt;
 use repr::{ColumnName, Datum, RelationDesc, RelationType, Row};
@@ -645,10 +647,18 @@ where
                 sql,
                 raw_plan,
                 decorrelated_plan,
+                row_set_finishing,
                 stage,
                 options,
             } => tx.send(
-                self.sequence_explain_plan(sql, raw_plan, decorrelated_plan, stage, options),
+                self.sequence_explain_plan(
+                    sql,
+                    raw_plan,
+                    decorrelated_plan,
+                    row_set_finishing,
+                    stage,
+                    options,
+                ),
                 session,
             ),
 
@@ -1239,6 +1249,7 @@ where
         sql: String,
         raw_plan: sql::RelationExpr,
         decorrelated_plan: expr::RelationExpr,
+        row_set_finishing: RowSetFinishing,
         stage: ExplainStage,
         options: ExplainOptions,
     ) -> Result<ExecuteResponse, failure::Error> {
@@ -1246,6 +1257,7 @@ where
             ExplainStage::Sql => sql,
             ExplainStage::RawPlan => {
                 let mut explanation = raw_plan.explain(&self.catalog);
+                explanation.explain_row_set_finishing(row_set_finishing);
                 if options.typed {
                     // TODO(jamii) does this fail?
                     explanation.explain_types(&BTreeMap::new());
@@ -1254,6 +1266,7 @@ where
             }
             ExplainStage::DecorrelatedPlan => {
                 let mut explanation = decorrelated_plan.explain(&self.catalog);
+                explanation.explain_row_set_finishing(row_set_finishing);
                 if options.typed {
                     explanation.explain_types();
                 }
@@ -1269,6 +1282,7 @@ where
                     .optimize(decorrelated_plan, self.catalog.indexes(), &eval_env)?
                     .into_inner();
                 let mut explanation = optimized_plan.explain(&self.catalog);
+                explanation.explain_row_set_finishing(row_set_finishing);
                 if options.typed {
                     explanation.explain_types();
                 }
