+++ src/expr/Cargo.toml
@@ -20,6 +20,7 @@ indexmap = "1.3.2"
 num_enum = "0.4.2"
 ordered-float = { version = "1.0.2", features = ["serde"] }
 ore = { path = "../ore" }
+pdqselect = "0.1.0"
 regex = "1.3.5"
 regex-syntax = "0.6.15"
 repr = { path = "../repr" }
