+++ src/expr/explain.rs
@@ -20,7 +20,7 @@
 //! It's important to avoid trailing whitespace everywhere, because it plays havoc with SLT
 use super::{
     AggregateExpr, EvalError, Id, IdHumanizer, JoinImplementation, LocalId, RelationExpr,
-    ScalarExpr,
+    RowSetFinishing, ScalarExpr,
 };
 use repr::RelationType;
 use std::collections::HashMap;
@@ -29,6 +29,8 @@ use std::collections::HashMap;
 pub struct Explanation<'a> {
     /// One ExplanationNode for each RelationExpr in the plan, in left-to-right post-order
     pub nodes: Vec<ExplanationNode<'a>>,
+    /// Anything else we want to mention at the end
+    pub appendix: String,
 }
 
 #[derive(Debug)]
@@ -62,11 +64,15 @@ impl<'a> std::fmt::Display for Explanation<'a> {
                 continue;
             }
 
-            // explain output shows up in SLT where the linter will not allow trailing whitespace, so trim stuff
+            // explain output shows up in SLT where the linter will not allow trailing whitespace, so need to trim stuff
             writeln!(f, "| {}", node.pretty.trim())?;
             for annotation in &node.annotations {
                 writeln!(f, "| | {}", annotation.trim())?;
             }
+        }
+
+        if !self.appendix.is_empty() {
+            writeln!(f, "\n{}", self.appendix.trim())?;
         }
 
         Ok(())
@@ -308,7 +314,10 @@ impl RelationExpr {
             }
         }
 
-        Explanation { nodes }
+        Explanation {
+            nodes,
+            appendix: String::new(),
+        }
     }
 }
 
@@ -323,6 +332,25 @@ impl<'a> Explanation<'a> {
                 "keys = ({})",
                 Separated(", ", keys.iter().map(|key| Indices(key)).collect())
             ));
+        }
+    }
+
+    pub fn explain_row_set_finishing(&mut self, row_set_finishing: RowSetFinishing) {
+        if !row_set_finishing.is_trivial() {
+            self.appendix.push_str(&format!(
+                "Finish order_by={} limit={} offset={} project={}",
+                Bracketed(
+                    "(",
+                    ")",
+                    Separated(", ", row_set_finishing.order_by.clone())
+                ),
+                match row_set_finishing.limit {
+                    Some(limit) => limit.to_string(),
+                    None => "none".to_owned(),
+                },
+                row_set_finishing.offset,
+                Bracketed("(", ")", Indices(&row_set_finishing.project))
+            ))
         }
     }
 }
