+++ src/sql/explain.rs
@@ -21,7 +21,7 @@
 //! It's important to avoid trailing whitespace everywhere, because it plays havoc with SLT
 
 use crate::expr::{AggregateExpr, JoinKind, RelationExpr, ScalarExpr};
-use expr::{Id, IdHumanizer};
+use expr::{Id, IdHumanizer, RowSetFinishing};
 use repr::{RelationType, ScalarType};
 use std::collections::{BTreeMap, HashMap};
 
@@ -29,6 +29,8 @@ use std::collections::{BTreeMap, HashMap};
 pub struct Explanation<'a> {
     /// One ExplanationNode for each RelationExpr in the plan, in left-to-right post-order
     pub nodes: Vec<ExplanationNode<'a>>,
+    /// Anything else we want to mention at the end
+    pub appendix: String,
 }
 
 #[derive(Debug)]
@@ -73,6 +75,10 @@ impl<'a> std::fmt::Display for Explanation<'a> {
                     }
                 }
             }
+        }
+
+        if !self.appendix.is_empty() {
+            writeln!(f, "\n{}", self.appendix.trim())?;
         }
 
         Ok(())
@@ -324,7 +330,10 @@ impl RelationExpr {
             }
         }
 
-        Explanation { nodes }
+        Explanation {
+            nodes,
+            appendix: String::new(),
+        }
     }
 }
 
@@ -339,6 +348,25 @@ impl<'a> Explanation<'a> {
                 "keys = ({})",
                 Separated(", ", keys.iter().map(|key| Indices(key)).collect())
             ));
+        }
+    }
+
+    pub fn explain_row_set_finishing(&mut self, row_set_finishing: RowSetFinishing) {
+        if !row_set_finishing.is_trivial() {
+            self.appendix.push_str(&format!(
+                "Finish order_by={} limit={} offset={} project={}",
+                Bracketed(
+                    "(",
+                    ")",
+                    Separated(", ", row_set_finishing.order_by.clone())
+                ),
+                match row_set_finishing.limit {
+                    Some(limit) => limit.to_string(),
+                    None => "none".to_owned(),
+                },
+                row_set_finishing.offset,
+                Bracketed("(", ")", Indices(&row_set_finishing.project))
+            ))
         }
     }
 }
